+++ tests/integration/test_fim/test_files/test_file_limit/test_file_limit_capacity_alerts.py
@@ -79,12 +79,14 @@ import sys
 
 import pytest
 from wazuh_testing import global_parameters
-from wazuh_testing.fim import LOG_FILE_PATH, callback_file_limit_capacity, generate_params, create_file, \
-    check_time_travel, REGULAR, delete_file, callback_file_limit_back_to_normal, \
-    callback_entries_path_count
+from wazuh_testing.fim import (LOG_FILE_PATH, generate_params, create_file, REGULAR, delete_file, wait_for_scheduled_scan)
 from wazuh_testing.tools import PREFIX
-from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
-from wazuh_testing.tools.monitoring import FileMonitor
+from wazuh_testing.tools.configuration import load_wazuh_configurations
+from wazuh_testing.tools.monitoring import FileMonitor, generate_monitoring_callback
+from wazuh_testing.fim_module import (CB_FILE_LIMIT_CAPACITY, ERR_MSG_DATABASE_PERCENTAGE_FULL_ALERT,
+    ERR_MSG_WRONG_CAPACITY_LOG_DB_LIMIT, ERR_MSG_WRONG_NUMBER_OF_ENTRIES, ERR_MSG_WRONG_INODE_PATH_COUNT,
+    CB_FILE_LIMIT_BACK_TO_NORMAL, ERR_MSG_DB_BACK_TO_NORMAL, ERR_MSG_FIM_INODE_ENTRIES)
+from wazuh_testing.fim_module.event_monitor import callback_entries_path_count
 
 # Marks
 
@@ -98,17 +100,18 @@ wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
 configurations_path = os.path.join(test_data_path, 'wazuh_conf.yaml')
 testdir1 = test_directories[0]
+scan_delay = 10
 
 # Configurations
 
 
 file_limit_list = ['100']
-conf_params = {'TEST_DIRECTORIES': testdir1, 'MODULE_NAME': __name__}
+conf_params = {'TEST_DIRECTORIES': testdir1}
 
-p, m = generate_params(extra_params=conf_params, modes=['scheduled'],
+params, metadata = generate_params(extra_params=conf_params, modes=['scheduled'],
                        apply_to_all=({'FILE_LIMIT': file_limit_elem} for file_limit_elem in file_limit_list))
 
-configurations = load_wazuh_configurations(configurations_path, __name__, params=p, metadata=m)
+configurations = load_wazuh_configurations(configurations_path, __name__, params=params, metadata=metadata)
 
 
 # Fixtures
@@ -123,13 +126,9 @@ def get_configuration(request):
 # Tests
 
 
-@pytest.mark.parametrize('percentage,tags_to_apply', [
-    (80, {'file_limit_conf'}),
-    (90, {'file_limit_conf'}),
-    (0, {'file_limit_conf'})
-])
-def test_file_limit_capacity_alert(percentage, tags_to_apply, get_configuration, configure_environment,
-                                   restart_syscheckd, wait_for_fim_start):
+@pytest.mark.parametrize('percentage', [(80), (90), (0)])
+def test_file_limit_capacity_alert(percentage, get_configuration, configure_environment, restart_syscheckd,
+                                   wait_for_fim_start):
     '''
     description: Check if the 'wazuh-syscheckd' daemon generates events for different capacity thresholds limits when
                  using the 'schedule' monitoring mode. For this purpose, the test will monitor a directory in which
@@ -144,9 +143,6 @@ def test_file_limit_capacity_alert(percentage, tags_to_apply, get_configuration,
         - percentage:
             type: int
             brief: Percentage of testing files to be created.
-        - tags_to_apply:
-            type: set
-            brief: Run test if matches with a configuration identifier, skip otherwise.
         - get_configuration:
             type: fixture
             brief: Get configurations from the module.
@@ -155,7 +151,7 @@ def test_file_limit_capacity_alert(percentage, tags_to_apply, get_configuration,
             brief: Configure a custom environment for testing.
         - restart_syscheckd:
             type: fixture
-            brief: Clear the 'ossec.log' file and start a new monitor.
+            brief: Clear the Wazuh logs file and start a new monitor.
         - wait_for_fim_start:
             type: fixture
             brief: Wait for realtime start, whodata start, or end of initial FIM scan.
@@ -178,59 +174,43 @@ def test_file_limit_capacity_alert(percentage, tags_to_apply, get_configuration,
 
     tags:
         - scheduled
-        - time_travel
     '''
-    check_apply_test(tags_to_apply, get_configuration['tags'])
 
     NUM_FILES = percentage + 1
 
     if percentage == 0:
         NUM_FILES = 0
-
+        # Create files up to desired database percentage to generate alerts
     if percentage >= 80:  # Percentages 80 and 90
         for i in range(NUM_FILES):
             create_file(REGULAR, testdir1, f'test{i}')
+        #Delete files to empty DB and return it to normal levels
     else:  # Database back to normal
         for i in range(91):
             delete_file(testdir1, f'test{i}')
 
-    check_time_travel(True, monitor=wazuh_log_monitor)
-
-    if percentage >= 80:  # Percentages 80 and 90
+    wait_for_scheduled_scan(True, interval=scan_delay, monitor=wazuh_log_monitor)
+    #Look for file_limit percentage alert configure value and check it matches with the expected percentage
+    if percentage >= 80:  
         file_limit_capacity = wazuh_log_monitor.start(timeout=global_parameters.default_timeout,
-                                                      callback=callback_file_limit_capacity,
-                                                      error_message='Did not receive expected '
-                                                                    '"DEBUG: ...: Sending DB ...% full alert." event'
-                                                      ).result()
+                                                      callback=generate_monitoring_callback(CB_FILE_LIMIT_CAPACITY),
+                                                      error_message=ERR_MSG_DATABASE_PERCENTAGE_FULL_ALERT).result()
 
-        if file_limit_capacity:
-            assert file_limit_capacity == str(percentage), 'Wrong capacity log for DB file_limit'
-        else:
-            raise AssertionError('Wrong capacity log for DB file_limit')
-    else:  # Database back to normal
+        assert file_limit_capacity == str(percentage), ERR_MSG_WRONG_CAPACITY_LOG_DB_LIMIT
+    # Check the is back on normal levels
+    else:  
         event_found = wazuh_log_monitor.start(timeout=global_parameters.default_timeout,
-                                              callback=callback_file_limit_back_to_normal,
-                                              error_message='Did not receive expected '
-                                                            '"DEBUG: ...: Sending DB back to normal alert." event'
-                                              ).result()
+                                              callback=generate_monitoring_callback(CB_FILE_LIMIT_BACK_TO_NORMAL),
+                                              error_message=ERR_MSG_DB_BACK_TO_NORMAL).result()
 
-        assert event_found, 'Event "Sending DB back to normal alert." not found'
-
+    # Get entries and path counts and check they match the expected values
     entries, path_count = wazuh_log_monitor.start(timeout=global_parameters.default_timeout,
                                                   callback=callback_entries_path_count,
-                                                  error_message='Did not receive expected '
-                                                                '"Fim inode entries: ..." event'
-                                                  ).result()
+                                                  error_message=ERR_MSG_FIM_INODE_ENTRIES).result()
 
-    check_time_travel(True, monitor=wazuh_log_monitor)
+    wait_for_scheduled_scan(True, interval=scan_delay, monitor=wazuh_log_monitor)
 
     if sys.platform != 'win32':
-        if entries and path_count:
-            assert entries == str(NUM_FILES) and path_count == str(NUM_FILES), 'Wrong number of inodes and path count'
-        else:
-            raise AssertionError('Wrong number of inodes and path count')
+        assert entries == str(NUM_FILES) and path_count == str(NUM_FILES), ERR_MSG_WRONG_INODE_PATH_COUNT
     else:
-        if entries:
-            assert entries == str(NUM_FILES), 'Wrong number of entries count'
-        else:
-            raise AssertionError('Wrong number of entries count')
+        assert entries == str(NUM_FILES), ERR_MSG_WRONG_NUMBER_OF_ENTRIES