+++ tests/integration/test_fim/test_files/test_file_limit/test_file_limit_full.py
@@ -79,11 +79,14 @@ import sys
 
 import pytest
 from wazuh_testing import global_parameters
-from wazuh_testing.fim import LOG_FILE_PATH, callback_file_limit_capacity, generate_params, create_file, REGULAR, \
-    callback_file_limit_full_database, callback_entries_path_count
+from wazuh_testing.fim import LOG_FILE_PATH, generate_params, create_file, REGULAR
 from wazuh_testing.tools import PREFIX
-from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
-from wazuh_testing.tools.monitoring import FileMonitor
+from wazuh_testing.tools.configuration import load_wazuh_configurations
+from wazuh_testing.tools.monitoring import FileMonitor, generate_monitoring_callback
+from wazuh_testing.fim_module import(CB_FILE_LIMIT_CAPACITY, ERR_MSG_DATABASE_FULL_ALERT_EVENT,
+    ERR_MSG_WRONG_VALUE_FOR_DATABASE_FULL, CB_DATABASE_FULL_COULD_NOT_INSERT, ERR_MSG_DATABASE_FULL_COULD_NOT_INSERT,
+    ERR_MSG_FIM_INODE_ENTRIES, ERR_MSG_WRONG_INODE_PATH_COUNT, ERR_MSG_WRONG_NUMBER_OF_ENTRIES)
+from wazuh_testing.fim_module.event_monitor import callback_entries_path_count
 
 # Marks
 
@@ -91,23 +94,23 @@ pytestmark = [pytest.mark.tier(level=1)]
 
 # Variables
 test_directories = [os.path.join(PREFIX, 'testdir1')]
-
 directory_str = ','.join(test_directories)
 wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
 configurations_path = os.path.join(test_data_path, 'wazuh_conf.yaml')
 testdir1 = test_directories[0]
 NUM_FILES = 10
+monitor_timeout = 40
 
 # Configurations
 
 file_limit_list = ['10']
-conf_params = {'TEST_DIRECTORIES': testdir1, 'MODULE_NAME': __name__}
+conf_params = {'TEST_DIRECTORIES': testdir1}
 
-p, m = generate_params(extra_params=conf_params,
+params, metadata = generate_params(extra_params=conf_params,
                        apply_to_all=({'FILE_LIMIT': file_limit_elem} for file_limit_elem in file_limit_list))
 
-configurations = load_wazuh_configurations(configurations_path, __name__, params=p, metadata=m)
+configurations = load_wazuh_configurations(configurations_path, __name__, params=params, metadata=metadata)
 
 
 # Fixtures
@@ -130,12 +133,7 @@ def extra_configuration_before_yield():
 
 # Tests
 
-
-@pytest.mark.parametrize('tags_to_apply', [
-    {'file_limit_conf'}
-])
-@pytest.mark.skip(reason="It will be blocked by wazuh/wazuh#9298, when it was solve we can enable again this test")
-def test_file_limit_full(tags_to_apply, get_configuration, configure_environment, restart_syscheckd):
+def test_file_limit_full( get_configuration, configure_environment, restart_syscheckd):
     '''
     description: Check if the 'wazuh-syscheckd' daemon generates proper events while the FIM database is in
                  'full database alert' mode for reaching the limit of files to monitor set in the 'file_limit' tag.
@@ -158,7 +156,7 @@ def test_file_limit_full(tags_to_apply, get_configuration, configure_environment
             brief: Configure a custom environment for testing.
         - restart_syscheckd:
             type: fixture
-            brief: Clear the 'ossec.log' file and start a new monitor.
+            brief: Clear the Wazuh logs file and start a new monitor.
 
     assertions:
         - Verify that the FIM database is in 'full database alert' mode
@@ -177,37 +175,29 @@ def test_file_limit_full(tags_to_apply, get_configuration, configure_environment
 
     tags:
         - scheduled
-        - time_travel
+        - whodata
+        - realtime
     '''
-    check_apply_test(tags_to_apply, get_configuration['tags'])
-
+    #Check that database is full and assert database usage percentage is 100%
     database_state = wazuh_log_monitor.start(timeout=global_parameters.default_timeout,
-                                             callback=callback_file_limit_capacity,
-                                             error_message='Did not receive expected '
-                                                           '"DEBUG: ...: Sending DB 100% full alert." event').result()
-
-    if database_state:
-        assert database_state == '100', 'Wrong value for full database alert'
-
+                                             callback=generate_monitoring_callback(CB_FILE_LIMIT_CAPACITY),
+                                             error_message=ERR_MSG_DATABASE_FULL_ALERT_EVENT).result()
+    assert database_state == '100', ERR_MSG_WRONG_VALUE_FOR_DATABASE_FULL
+    
+    # Create a file with the database being full - Should not generate events
     create_file(REGULAR, testdir1, 'file_full', content='content')
 
-    wazuh_log_monitor.start(
-        timeout=40, callback=callback_file_limit_full_database,
-        error_message='Did not receive expected '
-        '"DEBUG: ...: Couldn\'t insert \'...\' entry into DB. The DB is full, ..." event')
+    # Check new file could not be added to DB
+    wazuh_log_monitor.start(timeout=monitor_timeout, callback=generate_monitoring_callback(CB_DATABASE_FULL_COULD_NOT_INSERT),
+                            error_message=ERR_MSG_DATABASE_FULL_COULD_NOT_INSERT)
 
-    entries, path_count = wazuh_log_monitor.start(timeout=40, callback=callback_entries_path_count,
-                                                  error_message='Did not receive expected '
-                                                                '"Fim inode entries: ..., path count: ..." event'
-                                                  ).result()
+    # Check number of entries and paths in DB and assert the value matches the expected count
+    entries, path_count = wazuh_log_monitor.start(timeout=monitor_timeout, callback=callback_entries_path_count,
+                                                  error_message=ERR_MSG_FIM_INODE_ENTRIES).result()
 
     if sys.platform != 'win32':
         if entries and path_count:
-            assert entries == str(NUM_FILES) and path_count == str(NUM_FILES), 'Wrong number of inodes and path count'
-        else:
-            raise AssertionError('Wrong number of inodes and path count')
+            assert entries == str(NUM_FILES) and path_count == str(NUM_FILES), ERR_MSG_WRONG_INODE_PATH_COUNT
     else:
         if entries:
-            assert entries == str(NUM_FILES), 'Wrong number of entries count'
-        else:
-            raise AssertionError('Wrong number of entries count')
+            assert entries == str(NUM_FILES), ERR_MSG_WRONG_NUMBER_OF_ENTRIES